+++ src/ts_backend_check/test_project/backend/models.py
@@ -29,6 +29,8 @@ class UserModel(models.Model):
 
     id = models.CharField(max_length=32)
     name = models.CharField(max_length=50)
+    # tsbc: inherit email (blank=True)
+    # ts-backend-check: from AbstractUser inherit password
 
 
 class BackendOnlyModel(models.Model):